+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 
+* modernize video player preview with `movi-player` (supports x264, x265, multiple audio tracks, subtitles)
+* update `marked` to 9.1.6 (fixes security vulnerabilities)
+* add cross-origin isolation headers to page templates
+* configure Babel build script to support private class properties and methods in `node_modules`
+
+## v0.30.0
+
 * now require PHP 7.0.0+
 * fix archive-single-item problem
 * add header/footer search stop condition